+++ orca/api/__init__.py
@@ -1,14 +1,13 @@
-from flask import Blueprint
-from flask_restx import Api
-
-from orca.api.apis.v1 import graph as graph_ns
-from orca.api.apis.v1 import ingestor as ingestor_ns
-from orca.api.apis.v1 import alert as alert_ns
-
-def initialize(graph):
-    blueprint = Blueprint('api', __name__, url_prefix='/v1')
-    api = Api(blueprint, title='ORCA API', version='0.1',)
-    api.add_namespace(graph_ns.initialize(graph))
-    api.add_namespace(ingestor_ns.initialize(graph))
-    api.add_namespace(alert_ns.initialize(graph))
-    return blueprint
+# Copyright 2020 OpenRCA Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
