+++ orca/api/manager.py
@@ -19,7 +19,7 @@ import cotyledon
 import flask
 from flask_cors import CORS
 
-from orca import api
+from orca.api import apiv1
 from orca.graph import graph
 
 
@@ -54,6 +54,6 @@ class APIService(cotyledon.Service):
 
     def _initialize_application(self):
         app = flask.Flask(__name__)
-        app.register_blueprint(api.initialize(self._graph))
+        app.register_blueprint(apiv1.initialize(self._graph))
         CORS(app)
         return app
